requirements.txt
aiohttp==3.6.2
altair==4.1.0
altgraph==0.17
astor==0.8.1
async-generator==1.10
async-timeout==3.0.1
auto-py-to-exe==2.7.8
backcall==0.2.0
base58==2.0.1
blinker==1.4
boto3==1.15.11
botocore==1.18.11
bottle==0.12.18
bottle-websocket==0.2.9
brotlipy==0.7.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
docopt==0.6.2
Eel==0.12.4
entrypoints==0.3
enum-compat==0.0.3
future==0.18.2
gast==0.2.2
gevent==20.9.0
gevent-websocket==0.10.1
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
greenlet==0.4.17
grpcio==1.4.0
h5py==2.10.0
idna-ssl==1.1.0
importlib-metadata==2.0.0
ipykernel==4.5.1
ipython-genutils==0.2.0
ipywidgets==4.0.1
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.17.0
json5==0.9.5
kiwisolver==1.2.0
Mako==1.1.3
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==2.0.1
mistune==0.8.4
numpy==1.8.2
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.1.0
packaging==20.4
pandocfilters==1.4.2
parso==0.7.0
pathtools==0.1.2
pefile==2019.4.18
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.8.0
pyarrow==1.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.9
PyJWT==1.7.1
PySocks==1.7.1
python-dateutil==2.8.1
PyYAML==5.3.1
pyzmq==19.0.2
requests==2.17.3
requests-oauthlib==1.3.0
s3transfer==0.3.3
Send2Trash==1.5.0
six==1.15.0
streamlit==0.67.1
tensorflow==2.1.0
tornado==6.0.4
traitlets==4.3.3
tzlocal==2.1
urllib3==1.2.2
validators==0.18.1
watchdog==0.10.3
webencodings==0.5.1
Werkzeug==0.16.1
whichcraft==0.6.1
wincertstore==0.2
wrapt==1.12.1
yarg==0.1.9
zipp==3.1.0
zope.event==4.5.0
zope.interface==5.1.2
